+++ tensorboard/compat/BUILD
@@ -15,6 +15,7 @@ py_library(
     srcs_version = "PY2AND3",
     visibility = ["//visibility:public"],
     deps = [
+        "//tensorboard:expect_tensorflow_installed",
         "//tensorboard/compat/tensorflow_stub",
     ],
 )
